% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{add_document_table}
\alias{add_document_table}
\title{Data.frame of "other recovery documents" tables from ECOS}
\format{A data frame with 3136 rows and 6 variables
\itemize{
  \item{Date}{The document date as recorded in ECOS}
  \item{Citation Page}{The Federal Register citation for the document}
  \item{Title}{The text of the document link, i.e., the \code{<a>} tag}
  \item{Document Type}{The type of document, e.g., 5-year review notice}
  \item{Doc_Link}{The URL of the document, i.e., the \code{href}; may be
    relative or absolute}
  \item{Species}{The ECOS-standard scientific name of the species}
}}
\source{
\link{http://ecos.fws.gov}
}
\usage{
add_document_table
}
\description{
ECOS \url{http://ecos.fws.gov} includes a table named "Other Recovery 
Documents" for most but not all species. This data.frame was collected by
scraping all of these tables from each listed species' page and \code{rbind}
the rows to a single df.
}
\keyword{datasets}

