% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_links}
\alias{all_links}
\title{Data.frame of all links for each species in ECOS}
\format{A data frame with 40815 rows and 4 variables
\itemize{
  \item{Scientific_Name}{The scientific name, as recorded in ECOS}
  \item{href}{The URL given on the ECOS page}
  \item{link}{The URL after completing with the correct prefix}
  \item{text}{The text of the link, i.e., the <a> tag}
}}
\source{
\link{http://ecos.fws.gov}
}
\usage{
all_links
}
\description{
Data.frame of all links for each species in ECOS
}
\keyword{datasets}

