% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_tables.R
\name{bind_tables}
\alias{bind_tables}
\title{Extract named tables from a list of tables extracted for species}
\usage{
bind_tables(ls, table)
}
\arguments{
\item{ls}{The list of tables from \link{get_tables}}
}
\value{
A data.frame of type specified by \code{table}
}
\description{
More detailed description of the function
}
\examples{
\dontrun{
dplyr::distinct(bind_tables(test, "SP_TAB"))
}
}
\seealso{
\link{get_tables}
}

