% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{bulk_species_links}
\alias{bulk_species_links}
\title{Gather all links for all or a subset of ECOS species.}
\usage{
bulk_species_links(data, subset = "All")
}
\arguments{
\item{subset}{A string describing which group of links to fetch; one of
\itemize{
  \item `All`
  \item `FWS Domestic`
  \item `FWS Foreign`
  \item `NMFS`
}}
}
\value{
A data.frame with links, species, etc.
}
\description{
I need a better description of the function...
}
\examples{
# one or more lines to demo the function
}

