% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_fname}
\alias{clean_fname}
\title{Clean a file name}
\usage{
clean_fname(f)
}
\arguments{
\item{f}{The character string of a file name}
}
\description{
Clean a file name
}
\examples{
\dontrun{
  clean_fname("this&that are (good).pdf")
}
}

