% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_pic}
\alias{download_pic}
\title{Download an image file}
\usage{
download_pic(url, subd, pause = TRUE)
}
\arguments{
\item{url}{The URL of the image to download}

\item{subd}{The directory to which the image will be written}
}
\description{
Download an image file
}
\details{
Just a wrapper around \link[httr]{GET} with several checks to make
downloads more likely / robust.
}

