% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_species_page}
\alias{download_species_page}
\title{Download a species' ECOS page}
\usage{
download_species_page(url, file, overwrite = FALSE)
}
\arguments{
\item{url}{The URL of the HTML page to download}

\item{file}{The file to which the HTML will be written}

\item{overwrite}{Whether to overwrite pre-existing file}
}
\value{
A data_frame with four variables: \describe{
  \item{url}{The URL that was downloaded or attempted}
  \item{dest}{The file to which the HTML was written}
  \item{success}{Whether the download was a \code{success} or \code{failed}}
  \item{htmlCheck}{Whether the downloaded file is an HTML [bool]}
}
}
\description{
Download a species' ECOS page
}
\details{
A wrapper around \link[httr]{GET} with several checks to make
downloads more likely / robust.
}
\examples{
\dontrun{
  get_species_url("Helianthus paradoxus") \%>\% 
    download_species_page("~/Downloads/HELPAR_2016-12-08.html")
}
}
