% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecos_doc_links}
\alias{ecos_doc_links}
\title{Data.frame of document links for each species in ECOS}
\format{A data frame with 23022 rows and 5 variables
\itemize{
  \item{Scientific_Name}{The scientific name, as recorded in ECOS}
  \item{href}{The URL given on the ECOS page}
  \item{link}{The URL after completing with the correct prefix}
  \item{text}{The text of the link, i.e., the <a> tag}
  \item{type}{The type of document, e.g., federal_register, recovery_plan}
}}
\source{
\link{http://ecos.fws.gov/tess_public}
}
\usage{
ecos_doc_links
}
\description{
Data.frame of document links for each species in ECOS
}
\keyword{datasets}

