% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosscraper-package.R
\docType{package}
\name{ecosscraper}
\alias{ecosscraper}
\alias{ecosscraper-package}
\title{ecosscraper: Get data and documents from ECOS}
\description{
ecosscraper: Get data and documents from ECOS
}
\section{About}{


The U.S. Fish and Wildlife Service provides documents (and some limited data
about the documents) on ECOS, \url{http://ecos.fws.gov}. But there is no API
for the documents, so we have to scrape the contents.
}

