% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_docs.R
\name{file_check}
\alias{file_check}
\title{Check if file exists and if so, if it is really a PDF}
\usage{
file_check(f)
}
\arguments{
\item{f}{File to check}
}
\value{
TRUE if file exists and is PDF, else FALSE
}
\description{
Check if file exists and if so, if it is really a PDF
}

