% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{fill_link}
\alias{fill_link}
\title{Fill relative links with appropriate prefix.}
\usage{
fill_link(x, base_ln)
}
\arguments{
\item{x}{A link (href) string; may be relative or absolute}

\item{base_ln}{A string indicating the base URL of current page}
}
\value{
The complete link, as available.
}
\description{
Fill relative links with appropriate prefix.
}
\examples{
# None at the moment...
}

