% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_listed}
\alias{filter_listed}
\title{Remove candidate and proposed species from a TECP data.frame}
\usage{
filter_listed(df)
}
\arguments{
\item{df}{A data.frame derived from \link{get_TECP_links}}

\item{patterns}{A list of additional patterns to be filtered out}
}
\value{
A data.frame containing only ESA-listed species
}
\description{
Remove candidate and proposed species from a TECP data.frame
}
\details{
Expects a data.frame from get_TECP_table. The df may have already 
passed through another filter (e.g., filter_domestic), but must contain the 
variable \code{Federal_Listing_Status}.
}

