% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Filter TECP_table to return only species from a given taxonomic group}
\usage{
filter_taxa(df, group = c())
}
\arguments{
\item{df}{A data.frame derived from \link{get_TECP_links}}

\item{group}{One or more of \itemize{
  \item{Amphibians}
  \item{Arachnids}
  \item{Birds}
  \item{Clams}
  \item{Conifers and Cycads}
  \item{Corals}
  \item{Crustaceans}
  \item{Ferns and Allies}
  \item{Fishes}
  \item{Flowering Plants}
  \item{Insects}
  \item{Lichens}
  \item{Mammals}
  \item{Reptiles}
  \item{Snails}
}}

\item{patterns}{A list of additional patterns to be filtered out}
}
\value{
A data.frame containing only species from the selected taxa
}
\description{
Filter TECP_table to return only species from a given taxonomic group
}
\details{
Expects a data.frame from get_TECP_table. The df may have already 
passed through another filter (e.g., filter_domestic), but must contain the 
variable \code{Species_Group}.
}
