% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_page.R
\name{find_phantom}
\alias{find_phantom}
\title{Find the phantomjs binary}
\usage{
find_phantom()
}
\description{
Find the phantomjs binary
}
\details{
This is taken directly from \link[webshot]{webshot}, where 
\code{find_phantomjs} is a non-exported function.
}

