% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_5yrev_links}
\alias{get_5yrev_links}
\title{Get the links for five-year review documents.}
\usage{
get_5yrev_links(df)
}
\arguments{
\item{df}{A data.frame of species' links.}
}
\value{
df, filtered for only five-year review links
}
\description{
For when we want to just get the reviews.
}
\examples{
\dontrun{
  # get_5yrev_links(all_links)
}
}

