% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_agmt_type_links}
\alias{get_agmt_type_links}
\title{Get a listing of link suffixes for HCPs, SHA, and CCA/As}
\usage{
get_agmt_type_links(url, type, verbose = TRUE)
}
\arguments{
\item{url}{The species' ECOS page URL to scrape}

\item{type}{The type of conservation agreement to search for; one of
\itemize{
  \item{HCP}
  \item{SHA}
  \item{CCA}
  \item{CCAA}
}}

\item{verbose}{Print messages while processing [default = TRUE]}
}
\value{
A data.frame with plan type, plan name, species, and link to the plan
}
\description{
Get a listing of link suffixes for HCPs, SHA, and CCA/As
}
\note{
Does not use the ECOS conservation plan page,
\url{},
because we know that many plans linked on species' ECOS pages do not 
appear in the conservation plan portal.
}
\examples{
\dontrun{
agmt <- get_species_url("Gila purpurea") \%>\%
          get_agmt_type_links(type = "HCP")
}
}

