% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_agreement_links}
\alias{get_agreement_links}
\title{Get a listing of link suffixes for HCPs, SHA, and CCA/As.}
\usage{
get_agreement_links()
}
\value{
A data.frame with region, type, and suffix
}
\description{
Uses the ECOS conservation plan page as the root and selects from dropdown.
}
\details{
THIS FUNCTION SHOULD BE DEPRECATED BECAUSE WE KNOW THE TESS CONSERVATION PLAN
PORTAL IS MISSING PLANS THAT ARE LINKED ON SPECIES' PAGES.
}
\examples{
\dontrun{
  agmt <- get_agreement_links()
}
}

