% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_tables.R
\name{get_cca_species}
\alias{get_cca_species}
\title{Get a table of species covered by CCA/As}
\usage{
get_cca_species(txt, url)
}
\arguments{
\item{txt}{The text to be parsed}

\item{url}{The url of the CCA/A plan summary data}
}
\description{
Get a table of species covered by CCA/As
}
