% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_cons_plan_links}
\alias{get_cons_plan_links}
\title{Get the links to conservation plan pages.}
\usage{
get_cons_plan_links(df)
}
\arguments{
\item{df}{A data.frame of species' links.}
}
\value{
df, filtered for links to conservation plans (SHA, HCP, CCAA)
}
\description{
Get the links to conservation plan pages.
}
\examples{
# get_cons_plan_links(all_links)
}

