% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_tables.R
\name{get_conservation_plan_data}
\alias{get_conservation_plan_data}
\title{Get data linked to a section 10 conservation plan}
\usage{
get_conservation_plan_data(url)
}
\arguments{
\item{url}{The URL of the summary page for a section 10 conservation plan}
}
\value{
A list of two items \describe{
  \item{table}{A data_frame of summary data about the agreement}
  \item{pdfs}{A data_frame of links to agreement documents}
}
}
\description{
Get data linked to a section 10 conservation plan
}
\details{
This function returns two distinct items, a table of summary data
and a table of linked documents. Most other functions in the package should
return a single item, but rather than visiting the conservation plan page 
twice, we get both pieces of content at the same time.
}
\examples{
\dontrun{
  url <- "https://ecos.fws.gov/conserv_plans/conservationPlan/plan?plan_id=1470"
  URARHO <- get_conservation_plan_table(url)
}
}

