% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_tables.R
\name{get_conservation_plan_land_use}
\alias{get_conservation_plan_land_use}
\title{Get table of land use types per conservation plan}
\usage{
get_conservation_plan_land_use(uses, plan_url)
}
\arguments{
\item{uses}{A character vector of covered species from a conservation plan's
data table, e.g., from get_conservation_plan_data}

\item{plan_url}{The url for the conservation plan summary page; needed for 
JOINing with other cons. plan datasets}
}
\value{
A table of land use by plan url, with one land use per row
}
\description{
Get table of land use types per conservation plan
}
