% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_tables.R
\name{get_conservation_plan_listed_species}
\alias{get_conservation_plan_listed_species}
\title{Get table of species covered by a conservation plan}
\usage{
get_conservation_plan_listed_species(spp, plan_url)
}
\arguments{
\item{spp}{A character vector of covered species from a conservation plan's
data table, e.g., from get_conservation_plan_data}

\item{plan_url}{The url for the conservation plan summary page; needed for 
JOINing with other cons. plan datasets}
}
\value{
A table of species by plan url, with one species per row
}
\description{
Get table of species covered by a conservation plan
}
