% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_tables.R
\name{get_conservation_plan_nonlisted_spp}
\alias{get_conservation_plan_nonlisted_spp}
\title{Get table of non-listed species covered by a conservation plan}
\usage{
get_conservation_plan_nonlisted_spp(spp, plan_url)
}
\arguments{
\item{spp}{A character vector of covered species from a conservation plan's
data table, e.g., from get_conservation_plan_data}

\item{plan_url}{The url for the conservation plan summary page; needed for 
JOINing with other cons. plan datasets}
}
\value{
A table of species by plan url, with one species per row
}
\description{
NOTE: we can't pull out just the scientific names of covered species because 
of variation in how FWS records scientific names in the conservation plan 
summary pages. \code{grepl} and other tools should be used to filter.
}
