% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_tables.R
\name{get_counties}
\alias{get_counties}
\title{Get the table of counties of occurrence from ECOS}
\usage{
get_counties(url)
}
\arguments{
\item{url}{The URL of the counties-of-occurrence page for a species}
}
\value{
A data_frame of counties of occurrence provided by ECOS
}
\description{
Get the table of counties of occurrence from ECOS
}
\examples{
\dontrun{
  url <- "https://ecos.fws.gov/ecp0/profile/countiesBySpecies?entityId=2"
  URARHO <- get_counties(url)
}
}

