% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_fedreg_links}
\alias{get_fedreg_links}
\title{Get the links for Federal Register documents.}
\usage{
get_fedreg_links(df)
}
\arguments{
\item{df}{A data.frame of species' links.}
}
\value{
df, filtered for only Fed. Reg. links
}
\description{
For when we want to just get the Fed. Reg. documents.
}
\examples{
# get_fedreg_links(all_links)
}

