% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_tables.R
\name{get_link_table}
\alias{get_link_table}
\title{Get a data.frame of links and their titles from a web page}
\usage{
get_link_table(pg)
}
\description{
Get a data.frame of links and their titles from a web page
}

