% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_recovery_links}
\alias{get_recovery_links}
\title{Get the links for recovery plans.}
\usage{
get_recovery_links(df)
}
\arguments{
\item{df}{A data.frame of species' links.}
}
\value{
df, filtered for only recovery plan links
}
\description{
For when we want to just get the recovery plans.
}
\examples{
# get_recovery_links(all_links)
}

