% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_recovery_table}
\alias{get_recovery_table}
\title{Get the table at https://ecos.fws.gov/ecp0/pub/speciesRecovery.jsp}
\usage{
get_recovery_table(subd)
}
\description{
Get the table at https://ecos.fws.gov/ecp0/pub/speciesRecovery.jsp
}

