% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_species_links}
\alias{get_species_links}
\title{Return all links on a species' ECOS page.}
\usage{
get_species_links(species, pause = TRUE)
}
\arguments{
\item{species}{Scientific name of a species, as given on ECOS}
}
\value{
A data.frame with
\describe{
  \item{Scientific_Name}{The species' scientific name, as given by ECOS}
  \item{href}{The link as given in the <a> tag on the ECOS page}
  \item{link}{The link, in absolute form, if available}
  \item{text}{The anchor text of the <a> tag, if available}}
}
\description{
Return a data.frame with links and their text, from a species' page on ECOS.
}
\examples{
# None at the moment...
}

