% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_species_page}
\alias{get_species_page}
\title{Return the ECOS species profile for a given species.}
\usage{
get_species_page(species)
}
\arguments{
\item{species}{The scientific name of a species, as given in ECOS}
}
\value{
A page from \code{rvest::read_html}
}
\description{
Each species in ECOS has its own profile page, which can be scraped; this
simply returns the profile page.
}
\details{
There are several \code{try} statements because of the persnickettiness of
ECOS and/or to guard against bad species names. Most errors are caught by
\link{get_species_url}.
}
\examples{
# one or more lines to demo the function
}

