% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_tables.R
\name{get_species_page_summary}
\alias{get_species_page_summary}
\title{Get a summary data.frame of an ECOS page scrape}
\usage{
get_species_page_summary(pg, url, species)
}
\arguments{
\item{pg}{An ECOS species page from \link{get_species_page}}

\item{url}{The url of the species page}

\item{species}{The scientific name of the species whose page was scraped}
}
\description{
Get a summary data.frame of an ECOS page scrape
}

