% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_tables.R
\name{get_species_tables}
\alias{get_species_tables}
\title{Get all tables from a species' ECOS page}
\usage{
get_species_tables(url)
}
\arguments{
\item{url}{The path to the species' ECOS page}
}
\value{
A list of tables, named per \link{get_table_type}, and one
  table (\code{scrape_info}) that records information about the scrape
}
\description{
Get all tables from a species' ECOS page
}
\examples{
get_species_tables(TECP_domestic$Species_Page[1])
}
\seealso{
\link{get_table}, \link{get_table_type}
}

