% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_species_tsn}
\alias{get_species_tsn}
\title{Return a species' TSN from their ECOS page}
\usage{
get_species_tsn(url)
}
\arguments{
\item{url}{The URL or path to a local HTML file for a species}
}
\description{
Return a species' TSN from their ECOS page
}
\details{
FWS uses at least four different keys for species, including the TSN
that is defined by ITIS (\url{http://itis.gov}). The TSN is used for some 
JSON data queries; this function simplifies extraction.
}
\examples{
\dontrun{
  url <- "https://ecos.fws.gov/ecp0/profile/speciesProfile?spcode=A001"
  get_species_tsn(url)
}
}

