% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_species_url}
\alias{get_species_url}
\title{Return the profile page on ECOS for a given species.}
\usage{
get_species_url(df, species)
}
\arguments{
\item{df}{A data.frame returned from get_TECP_table}

\item{species}{The scientific name of a species, as given by ECOS}
}
\value{
The URL of the species' ECOS profile
}
\description{
Uses the species code in ecos_listed_spp data.frame to form the URL.
}
\examples{
get_species_url("Ursus arctos horribilis")
}

