% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_tables.R
\name{get_table}
\alias{get_table}
\title{Return a table from an ECOS page}
\usage{
get_table(tab)
}
\arguments{
\item{tab}{A table from an rvest::html_nodes object}
}
\value{
The table as a data.frame
}
\description{
Return a table from an ECOS page
}
\seealso{
\link{get_tables}
}

