% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_tables.R
\name{get_table_type}
\alias{get_table_type}
\title{Return the category of ECOS table}
\usage{
get_table_type(df)
}
\arguments{
\item{df}{A data.frame from an ECOS table}
}
\value{
The category for the table; one of
  \itemize{
    \item{SP_TAB}{A table with basic listed species information}
    \item{FR_TAB}{A table with Federal Register documents}
    \item{CH_TAB}{A table with critical habitat documents}
    \item{REC_TAB}{A table with recovery plan information}
    \item{DOC_TAB}{A table with additional documents}
    \item{REV_TAB}{A table with 5-year review documents}
    \item{others}{One of several table types, e.g., HCP documents}
  }
}
\description{
Return the category of ECOS table
}
\details{
Each species' ECOS page contains several tables, but which tables
are present on any given page can vary. The tables are not given attr IDs, 
which would allow easy identification of the table types, so we use the 
headings of each table to determine the types.
}

