% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_tables.R
\name{get_tables}
\alias{get_tables}
\title{Wrapper to get all tables from a species' ECOS page}
\usage{
get_tables(sp)
}
\arguments{
\item{sp}{The path to the species' ECOS page}
}
\value{
A list of tables, named per \link{get_table_type}, and one
  table that records information about the scrape
}
\description{
More detailed description of the function
}
\examples{
get_tables(TECP_domestic$Species_Page[1])
}
\seealso{
\link{get_table}, \link{get_table_type}
}

