% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TECP_table.R
\name{get_TECP_table}
\alias{get_TECP_table}
\title{Get a df of threatened, endangered, candidate, and proposed species from ECOS}
\usage{
get_TECP_table()
}
\value{
A data.frame with 11 variables:
\itemize{
  \item Scientific_Name
  \item Common_Name
  \item Species_Code
  \item Critical_Habitat
  \item Species_Group
  \item Lead_Region
  \item Federal_Listing_Status
  \item Special_Rules
  \item U_S__or_ForeignListed
  \item Where_Listed
  \item Species_Page
}
}
\description{
Returns threatened, endangered, candidate, and proposed species from ECOS.
}
\details{
Uses the URL set in options()$TE_list to get a data.frame that 
includes links for fetching data for any species with pages on TESS. This 
data is loaded
}
\examples{
# To replace the TECP_table that is loaded on startup (attach)
TECP_table <- get_TECP_table()
head(TECP_table)
}
