% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HCP_SHA}
\alias{HCP_SHA}
\title{Summary data on Habitat Conservation Plans and Safe Harbor Agreements}
\format{A data frame with 1288 rows and 30 variables
\describe{
\item{\code{Link}}{Link to the local copy of the agreement page}
\item{\code{Outlink}}{Link(s) to any documents outside of the agreement page}
\item{\code{Plan_Name}}{Name of the agreement, as given on ECOS}
\item{\code{Agreement_Type}}{Type of agreement; one of HCP, SHA, CCA, CCAA}
\item{\code{USFWS_Regions}}{FWS regions in which the agreement occurs}
\item{\code{USFWS_Field_Offices}}{FWS offices managing the agreement}
\item{\code{Permits}}{FWS identifiers for CCAA permits}
\item{\code{Status}}{In-progress or Expired}
\item{\code{NEPA_Process}}{The type of NEPA analysis performed}
\item{\code{Involved_Agency}}{Agency(ies) involved in the agreement}
\item{\code{Budget_Type}}{Whether a Forest Plan or something else (NA)}
\item{\code{Date_Noticed_in_Federal_Register}}{Date of FR notice}
\item{\code{Date_Permit_Issued}}{Start date of the permit}
\item{\code{Date_Application_Withdrawn}}{Date permit application withdrawn}
\item{\code{Date_Permit_Expired}}{End date of the permit}
\item{\code{Date_Permit_Denied}}{Date the permit was denied}
\item{\code{Date_Permit_Suspended}}{Date the permit was suspended}
\item{\code{Date_Permit_Revoked}}{Date the permit was revoked by FWS}
\item{\code{Location}}{Location of the agreement, as given in ECOS}
\item{\code{Listed_Species}}{ESA-listed species covered by the HCP/SHA; free 
  text that needs parsing}
\item{\code{Non-Listed_Species}}{Non-listed species covered by the HCP/SHA; 
  free text that needs parsing}
\item{\code{Habitat}}{Habitats encompassed by the agreement, as given in ECOS}
\item{\code{States}}{States in which the agreement occurs}
\item{\code{Size_Total_Area_Covered_by_State}}{Area (acres, usually) covered 
  by the agreement; free text that needs to be parsed}
\item{\code{Enrolled Size_Total_Area_Covered_by_State}}{Area (acres, usually) 
  enrolled under the agreement; free text that needs to be parsed}
\item{\code{Applicant_Type}}{Category of applicant for the CCAA}
\item{\code{Land_Use}}{Land use(s) covered by the agreement, as given in ECOS}
\item{\code{Duration}}{Agreement duration, years and months}
\item{\code{Plan_Agreement_Documents}}{Any attached agreement documents}
\item{\code{Federal_Register_Documents}}{Any attached Federal Register documents}
}}
\source{
\url{http://ecos.fws.gov}
}
\usage{
HCP_SHA
}
\description{
Section 10(a)(1)(B) of the US Endangered Species Act sets the mechanism by
which non-federal entities - private landowners, state governments, and
others - receive authorization to 'take' listed species, or coverage for
certain types of activities involving candidate species, in return for
taking conservation actions for those species. The U.S. Fish and Wildlife
Service (FWS) keeps basic records about these agreements on their ECOS
website, \url{http://ecos.fws.gov}. This data is for Habitat Conservation
Plans (HCPs) and Safe Harbor Agreements (SHAs).
}
\keyword{datasets}

