% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_page.R
\name{install_phantom}
\alias{install_phantom}
\title{Install phantomjs}
\usage{
install_phantom()
}
\description{
Install phantomjs
}
\details{
A simple wrapper over \link[webshot]{install_phantomjs}. NOTE that 
this requires using the fork at \url{https://github.com/jacob-ogre/webshot}
if you are using R version 3.3.2, at least until the pull request is merged
into the main branch of \code{webshot}.
}
