% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_docs.R
\name{is_pdf}
\alias{is_pdf}
\title{Test if a file is a pdf}
\usage{
is_pdf(f)
}
\arguments{
\item{f}{Path to a file to test}
}
\value{
TRUE if pdftools::pdf_info thinks it's a PDF, else FALSE
}
\description{
Test if a file is a pdf
}

