% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{is_species_profile}
\alias{is_species_profile}
\title{Check if the requested page is an ECOS profile page.}
\usage{
is_species_profile(page)
}
\arguments{
\item{page}{An rvest read_html page}
}
\value{
logical; TRUE if a species profile, FALSE if "No species profile"
}
\description{
ECOS will return a page rather than 404 if the species code is wrong. This
checks that the page is not the "No species profile" page.
}
\examples{
# one or more lines to demo the function
}

