% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_tables.R
\name{join_for_links}
\alias{join_for_links}
\title{Join an ECOS table of text with the links to documents}
\usage{
join_for_links(tab, links, sp)
}
\arguments{
\item{tab}{The table (as a data.frame) to be joined if it has a Title var}

\item{links}{A data.frame with the Title of the link and the URL}

\item{sp}{The scientific name to be included in the returned data.frame}
}
\value{
A data.frame with URL, if tab includes a Title variable
}
\description{
Extracting tables with links in cells using rvest isn't straight-forward.
But by listing the links with \code{html_nodes(..., "href")} and the link 
text with \code{html_text(...)} (i.e., the title), then joining with the 
table Title field, we can associate every doc link with the other fields
in the table.
}

