% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_docs.R
\name{make_pdf_dest}
\alias{make_pdf_dest}
\title{Construct a path for a download PDF}
\usage{
make_pdf_dest(link, subd = "")
}
\arguments{
\item{link}{A URL from ECOS to download a PDF document}

\item{subd}{The subdirectory in which the download will be written}
}
\value{
The file path where the download will be written
}
\description{
Construct a path for a download PDF
}

