% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_base_dir}
\alias{set_base_dir}
\title{Set the base directory for ecosscraper downloads.}
\usage{
set_base_dir(path)
}
\arguments{
\item{path}{Path to base directory for downloads from ECOS}
}
\value{
Nothing
}
\description{
Creates the directory, recursively, if it doesn't exist.
}
\examples{
set_base_dir(path = "~/Downloads/test5")
}

