% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_TE_list_opt}
\alias{set_TE_list_opt}
\title{Set the URL of the TESS table}
\usage{
set_TE_list_opt(url)
}
\arguments{
\item{url}{The new URL from which the base TESS data is scraped}
}
\description{
Set the URL of the TESS table
}

