% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TECP_domestic}
\alias{TECP_domestic}
\title{Data.frame of base information for domestic ECOS species}
\format{A data frame with 1835 rows and 10 variables
\itemize{
  \item{Scientific_Name}{The scientific name, as recorded in ECOS}
  \item{Common_Name}{The common name, as recorded in ECOS}
  \item{Species_Code}{The four-character code assigned to species in ECOS}
  \item{Critical_Habitat}{CFR section under which CH was declared}
  \item{Species_Group}{Taxonomic group of species, as recorded in ECOS}
  \item{Lead_Region}{FWS region responsible for recovery}
  \item{Federal_Listing_Status}{At time of scraping}
  \item{Special_Rules}{CFR section under which any special rules were made}
  \item{Where_Listed}{Geographic extent of listed entity}
  \item{Species_Page}{URL dedicated to the species on ECOS}
}}
\source{
\link{http://ecos.fws.gov/tess_public}
}
\usage{
TECP_domestic
}
\description{
The data is the table is the \link{TECP} data filtered for domestic species.
}
\keyword{datasets}

