% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{bulk_species_links}
\alias{bulk_species_links}
\title{Gather all links for all or a subset of ECOS species.}
\usage{
bulk_species_links(data, ..., parallel = TRUE)
}
\arguments{
\item{data}{A data.frame with}

\item{...}{Zero or more filter conditions as used in \code{dplyr::filter}}

\item{parallel}{TRUE (default) to try parallel scraping}
}
\value{
A data.frame with links, species, etc.
}
\description{
Collects all of the links (<a href...>) from ECOS for species.
}
\details{
Automates the process of gathering links, which may in turn be followed or,
in the case of documents or tables, be downloaded.
}
\examples{
# one or more lines to demo the function
}
\seealso{
\link{remove_silly_links} \link{get_species_links}
}

