% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_domestic}
\alias{filter_domestic}
\title{Filter TECP_table to return only domestic ESA-listed species}
\usage{
filter_domestic(df)
}
\arguments{
\item{df}{A data.frame derived from \link{get_TECP_links}}

\item{patterns}{A list of additional patterns to be filtered out}
}
\value{
A data.frame containing only domestic species
}
\description{
Filter TECP_table to return only domestic ESA-listed species
}
\details{
Expects a data.frame from get_TECP_table. The df may have already 
passed through another filter (e.g., filter_listed), but must contain the 
variable \code{U_S__or_ForeignListed}.
}

