% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_5yrev_links}
\alias{get_5yrev_links}
\title{Get the links for ESA five-year review documents from ECOS species' pages}
\usage{
get_5yrev_links(df)
}
\arguments{
\item{df}{A data.frame of species' links.}
}
\value{
df, filtered for only five-year review links
}
\description{
Get the links for ESA five-year review documents from ECOS species' pages
}
\examples{
\dontrun{
  five_yr <- get_species_url("Gila purpurea") \%>\% 
               get_species_links() \%>\%
               get_5yrev_links()
}
}
