% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_tables.R
\name{get_conservation_plan_data}
\alias{get_conservation_plan_data}
\title{Get data linked to a section 10 conservation plan}
\usage{
get_conservation_plan_data(url, species)
}
\arguments{
\item{url}{The URL of the summary page for a section 10 conservation plan}

\item{species}{The species from whose ECOS page the plan URL was gathered}
}
\value{
A data_frame of summary data about the agreement or NULL
}
\description{
Get data linked to a section 10 conservation plan
}
\details{
The ECOS pages for conservation plans include covered species,
(\code{Listed_Species / Non-Listed_Species}), but those "lists" are just a
concatenation of common names, scientific names, where listed, etc. The
\code{species} parameter is used to tie outgoing links from a species' ECOS
page to the summary data. These can later be checked to make sure that all
covered species listed on the plan page are linked to the species' page.
}
\examples{
\dontrun{
  url <- "https://ecos.fws.gov/conserv_plans/conservationPlan/plan?plan_id=1470"
  URARHO <- get_conservation_plan_data(url, "Ursus arctos horribilis")
}
}

