% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_tables.R
\name{get_conservation_plan_doc_links}
\alias{get_conservation_plan_doc_links}
\title{Get links to documents on a section 10 conservation plan page}
\usage{
get_conservation_plan_doc_links(url, species)
}
\arguments{
\item{url}{The URL of the summary page for a section 10 conservation plan}

\item{species}{The species from whose ECOS page the plan URL was gathered}
}
\value{
A data_frame of links to agreement documents
}
\description{
Get links to documents on a section 10 conservation plan page
}
\examples{
\dontrun{
  url <- "https://ecos.fws.gov/conserv_plans/conservationPlan/plan?plan_id=1470"
  conplan_docs <- get_conservation_plan_doc_links(url, 
                                                  "Ursus arctos horribilis")
}
}

