% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_docs.R
\name{get_document}
\alias{get_document}
\title{GET a document from ECOS.}
\usage{
get_document(link, subd, pause = TRUE)
}
\arguments{
\item{link}{A URL from ECOS to download a document}

\item{subd}{Subdirectory to which the document will be downloaded}

\item{pause}{Whether to pause for 0.5-3 seconds during scraping}
}
\value{
A data.frame with destination and success
}
\description{
Simple function to download a file from ECOS.
}
\examples{
# one or more lines to demo the function
}

