% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_fedreg_links}
\alias{get_fedreg_links}
\title{Get the links for Federal Register documents on ECOS species' pages}
\usage{
get_fedreg_links(df)
}
\arguments{
\item{df}{A data.frame of links from \link{get_species_links}}
}
\value{
df, filtered for only Fed. Reg. links
}
\description{
Get the links for Federal Register documents on ECOS species' pages
}
\examples{
  fed_reg <- get_species_url("Gila purpurea") \%>\% 
               get_species_links() \%>\%
               get_fedreg_links()
}

