% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_recovery_links}
\alias{get_recovery_links}
\title{Get the links for recovery plans on ECOS species' pages}
\usage{
get_recovery_links(df)
}
\arguments{
\item{df}{A data.frame of species' links.}
}
\value{
df, filtered for only recovery plan links
}
\description{
Get the links for recovery plans on ECOS species' pages
}
\examples{
  recovery <- get_species_url("Gila purpurea") \%>\% 
                get_species_links() \%>\%
                get_recovery_links()
}

