% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_species_links}
\alias{get_species_links}
\title{Return all links on a species' ECOS page}
\usage{
get_species_links(url, clean = TRUE, pause = TRUE, verbose = TRUE)
}
\arguments{
\item{url}{The URL of a species' page on ECOS [default = NULL]}

\item{clean}{Remove useless links (e.g., www.usa.gov) [default = TRUE]}

\item{pause}{0.5-3 second pause to be nice to the server [default = TRUE]}

\item{verbose}{Message the species being processed [default = TRUE]}
}
\value{
A data.frame with variables:
\describe{
  \item{Scientific_Name}{The species' scientific name, as given by ECOS}
  \item{href}{The link as given on the ECOS page, if available}
  \item{link}{The link, in absolute form, if available}
  \item{text}{The anchor text of the <a> tag, if available}
}
}
\description{
Return a data.frame with links (and anchor text) from a species' ECOS page.
}
\note{
Either \code{url} or \code{page} must be specified.
}
\examples{
\dontrun{
  res <- get_species_url("Gila purpurea") \%>\% get_species_links()
  head(res)
}
}
\seealso{
\link{remove_silly_links} \link{get_bulk_species_links}
}

