% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_page.R
\name{get_species_page}
\alias{get_species_page}
\title{Return the ECOS page for a given species as an \link[xml2]{xml2} objects}
\usage{
get_species_page(url, verbose = FALSE)
}
\arguments{
\item{url}{The url for a species' ECOS profile}

\item{verbose}{Whether message that the page is being fetched [default = FALSE]}
}
\value{
A page returned by \code{rvest::read_html}
}
\description{
Return the ECOS page for a given species as an \link[xml2]{xml2} objects
}
\details{
There are several \code{try} statements because of the 
persnickettiness of #' ECOS. Most errors should caught by 
\link{get_species_url}.
}
\note{
Unlike most other functions in \link{ecosscraper}, \code{verbose} is
\code{FALSE} by default. This function is mostly a helper and other "primary"
functions are \code{verbose} by default.
}
\examples{
\dontrun{
  page <- get_species_url("Abies guatemalensis") \%>\% get_species_page()
}
}
