% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_species_tables}
\alias{get_species_tables}
\title{Get all tables from a species' ECOS page}
\usage{
get_species_tables(url, verbose = TRUE)
}
\arguments{
\item{url}{The path to the species' ECOS page}

\item{verbose}{Print a message about tables being fetched [default = TRUE]}
}
\value{
A list of tables, named per \link{get_table_type}, and one
  table (\code{scrape_info}) that records information about the scrape
}
\description{
Get all tables from a species' ECOS page
}
\examples{
\dontrun{
  tabs <- get_species_tables(TECP_domestic$Species_Page[1])
  tab2 <- get_species_url("Abies guatemalensis") \%>\% get_species_tables()
}
}
\seealso{
\link{get_table}, \link{get_table_type}
}

