% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_species_url}
\alias{get_species_url}
\title{Return the profile page on ECOS for a given species.}
\usage{
get_species_url(species)
}
\arguments{
\item{species}{The scientific name of a species, as given by ECOS}

\item{df}{A data.frame returned from get_TECP_table}
}
\value{
The URL of the species' ECOS profile
}
\description{
Return the profile page on ECOS for a given species.
}
\examples{
get_species_url("Gila purpurea")
}

