% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_TECP_table}
\alias{get_TECP_table}
\title{Get a table of T, E, C, and P data.}
\usage{
get_TECP_table()
}
\value{
A data.frame with variables:
\itemize{
  \item Scientific_Name
  \item Common_Name
  \item Species_Code
  \item Critical_Habitat
  \item Species_Group
  \item Lead_Region
  \item Federal_Listing_Status
  \item Special_Rules
  \item Where_Listed
  \item Species_Page
}
}
\description{
Use the URL set in options()$TE_list to get a data.frame that includes links
for fetching data for any species with pages on TESS.
}
\examples{
all_spp <- get_TECP_table()
head(all_spp)
}

